+++ controllers/contacts/index.js
@@ -1,0 +1,14 @@
+const getAll = require('./getAll')
+const getContactById = require('./getContactById')
+const addContact = require('./addContact')
+const updateContact = require('./updateContact')
+const removeContact = require('./removeContact')
+
+
+module.exports = {
+    getAll,
+    getContactById,
+    addContact,
+    updateContact,
+    removeContact,
+}
