+++ controllers/contacts/updateContact.js
@@ -1,0 +1,23 @@
+const contacts = require("../../models/contacts")
+const {RequestError} = require("../../helpers")
+const {updateSchema} = require('../../schema/contacts')
+
+const updateContact = async (req, res, next) => {
+    try {
+      const {error} = updateSchema.validate(req.body);
+        if(error){
+          throw RequestError(400, error.message)
+        }
+      const {contactId} = req.params;
+      const result = await contacts.updateContact(contactId, req.body);
+        if(!result){
+          throw RequestError(404, "Not found")
+        }
+      res.json(result)
+    } catch (error) {
+      next(error)
+    }
+    
+  }
+
+  module.exports = updateContact;
