+++ routes/api/contacts.js
@@ -2,91 +2,17 @@ const express = require('express')
 
 const router = express.Router()
 
-const contacts = require("../../models/contacts")
 
-const {RequestError} = require("../../helpers")
+const ctrl = require('../../controllers/contacts/index')
 
-const Joi = require('joi')
+router.get('/', ctrl.getAll)
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-})
+router.get('/:contactId', ctrl.getContactById)
 
-const updateSchema = Joi.object({
-  name: Joi.string(),
-  email:Joi.string(),
-  phone: Joi.string(),
-}).min(1)
+router.post('/', ctrl.addContact)
 
-router.get('/', async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error)
-  }
-})
+router.delete('/:contactId', ctrl.removeContact)
 
-router.get('/:contactId', async (req, res, next) => {
-  try {
-    const {contactId} = req.params;
-    const result = await contacts.getContactById(contactId);
-      if(!result){
-        throw RequestError(404, "Not found")
-      }
-    res.json(result)
-  } catch (error) {
-    next(error)
-  }
-  
-})
-
-router.post('/', async (req, res, next) => {
-  try {
-    const {error} = addSchema.validate(req.body);
-      if(error){
-        throw RequestError(400, error.message)
-      }
-    const result = await contacts.addContact(req.body);
-    res.status(201).json(result)
-  } catch (error) {
-    next(error)
-  }
-  
-})
-
-router.delete('/:contactId', async (req, res, next) => {
-  try {
-    const {contactId} = req.params;
-    const result = await contacts.removeContact(contactId);
-      if(!result){
-        throw RequestError(404, "Not found")
-      }
-    res.json({message: "Delete success"})
-  } catch (error) {
-    next(error)
-  }
-  
-})
-
-router.put('/:contactId', async (req, res, next) => {
-  try {
-    const {error} = updateSchema.validate(req.body);
-      if(error){
-        throw RequestError(400, error.message)
-      }
-    const {contactId} = req.params;
-    const result = await contacts.updateContact(contactId, req.body);
-      if(!result){
-        throw RequestError(404, "Not found")
-      }
-    res.json(result)
-  } catch (error) {
-    next(error)
-  }
-  
-})
+router.put('/:contactId', ctrl.updateContact)
 
 module.exports = router;
